Move Blosc2-related utilities into their own module
As many will be used both by services and clients.
